fix: handle size truncation on 32 bit systems
If tempfile is used on a 32 bit system, usize is u32. This means that
casting u64 to usize may truncate the value.

Test added to clarify the issue.

+++ src/spooled.rs
@@ -96,7 +96,7 @@ impl SpooledTempFile {
     }
 
     pub fn set_len(&mut self, size: u64) -> Result<(), io::Error> {
-        if size as usize > self.max_size {
+        if size > self.max_size as u64 {
             self.roll()?; // does nothing if already rolled over
         }
         match &mut self.inner {
